mysite/requirements.txt
asgiref==3.9.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==6.0
django-debug-toolbar==6.1.0
django-filter==25.2
djangorestframework==3.16.1
docutils==0.22.3
drf-spectacular==0.29.0
drf-yasg==1.21.11
Faker==38.0.0
flake8==7.3.0
flake8-docstrings==1.7.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
pytest==8.0.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
sentry-sdk==2.48.0
snowballstemmer==3.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
